+++ update.sh
@@ -19,3 +19,4 @@ cp -r "$WORKDIR/matugen/"* "$HOME/.local/matugen/"
 sudo ln -sf "$HOME/.local/SDG-DMS-BARS/bar-presets.sh" /usr/bin/dmsbars
 sudo ln -sf "$HOME/.local/SDG-MANGO-LAYOUTS/layout-switch.sh" /usr/bin/manlayout
 sudo ln -sf "$HOME/.local/SDG-MANGO-CONF/mangocli.sh" /usr/bin/mangoconf
+sudo ln -sf "$HOME/.local/SDG-MANGO-LAYOUTS/equalize-scroller.sh" /usr/bin/equalize-scroller